+++ R/get_varcov_sandwich.R
@@ -9,6 +9,12 @@
 ) {
   dots <- list(...)
 
+  # make sure we have a "matrix" class
+  if (inherits(vcov_fun, "Matrix") || inherits(vcov_fun, "dpoMatrix")) {
+    check_if_installed("Matrix")
+    vcov_fun <- as.matrix(vcov_fun)
+  }
+
   if (is.null(vcov_args)) {
     vcov_args <- list()
   }